requirements.txt
tensorflow==2.20.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.4.0
numpy==1.26.4
